model_serving/llama_code_review_app/requirements.txt
fastapi
uvicorn
onnxruntime-gpu
transformers
numpy
python-multipart

